+++ thehive/app/org/thp/thehive/services/ObservableTypeSrv.scala
@@ -60,5 +60,5 @@ class ObservableTypeIntegrityCheckOps(val db: Database, val service: ObservableT
       case _ => Success(())
     }
 
-  override def globalCheck(): Map[String, Long] = Map.empty
+  override def globalCheck(): Map[String, Int] = Map.empty
 }
